Drop output of failed ReqLo eval if one succeed
When multiple modules are listed on a prereq command, drop the output of
those modules that fails to load (by the requirement load automated
mechanism) to only keep the output of the module whose load succeed.

Introduces procedures and variables to hold reports until all
requirement choices are tried to be loaded, then release held output
(drop or print). Exception is made for debug output which is always
immediately printed.

Fixes #254.