Rely on cmdModuleUnload error catch mech for prereq report
When attempting to unload a module, if this module has dependent loaded
it cannot be unloaded if auto_handling mode is disabled. Treat that
situation through the cmdModuleUnload error catch rather just with a
reportError call.

In case this situation happens through a 'module unload' command in a
modulefile evaluated in load mode, no error return code should be set.
However in other situations, cmdModuleUnload command should return an
error code.

So this change fixes 'switch' sub-command when the switched-off module
cannot be unloaded as it has some dependent module loaded. Now it fails
the whole switch process and no load of the switched-on module is not
tried.

Fixes #251.