More appropriate warn/err msg for failed ReqLo
When a module requirement is seen missing but the load of this module
was attempted, report a more specific error or warning message to let
user understand that the load of the requirement was attempted but
failed.

Thus distinction is made between missing requirements where a load
tentative was made and those were the no load attempt was performed
(when the automated module handling mode is disabled for instance).

Fixes #257.