Also register unload evaluation attempt
In addition to load evaluation attempt, also register the various unload
evaluation attempt to get a more appropriate warning or error message
regarding the missing requirement during a module load. When these
requirements are unloaded due to a conflict or a dependent unload/reload
mechanism, it seems better not to advise user to load this requirement
(as the dependency management system has unloaded it).