Report loaded conflict on corresp. module msg block
When loading a module, report any loaded conflict on the message
reporting block corresponding to this module load. This warning or error
message comes in addition to the eventual conflict unload message
reported under the message block of the main evaluation.

When a module load is forced and a conflict of this module is loaded,
report this conflict with the same concise message (no hint) whether an
unload of this conflicting module was attempted or not.

Fixes #261.