Distinguish DepRe load/unload phases in msgs
Adapt warning, error and info messages relative to the Dependent Reload
mechanism to distinguish the unload phase from the load (reload) phase
of this mechanism. In the automated module handling summary report,
unloaded modules via this mechanism are reported in the *Unloading
dependent* list and modules reloaded afterward are reported against the
"Reloading dependent* list.

Fixes #263.