Inhibit 'module load' in DepRe context
When reloading modules through the *Dependent Reload* automated
mechanism, disable the automated loading of modules by other modules with
the 'module load' modulefile command as done for the 'prereq' command.

It saves the mechanism from the reload of modules that have just been
unloaded (and asked by the user to be unloaded from instance when
performing a 'switch' command).

Fixes #265.