Fix colors init against badly formed MODULES_COLORS
Seems that 'array set' could partially set values to an array in case
the list passed to it is badly formed (odd number of element). This
situation has been witnessed on FreeBSD. So unset the array variable to
ensure it will not be found set at all by the remaining code of the
initColors procedure.