Control whether an implicit default should be selected
Add the ability to configure whether an implicit default version should
be defined for modules with no default version explicitly defined. When
enabled, which stays the default behavior, a module version is
automatically selected (latest one) when the generic name of the module
is passed. When implicit default selection is disabled, the name of
modules to evaluate should be fully qualified elsewhere an error is
returned. This option is set at `./configure` time with the
`--enable-implicit-default` and `--disable-implicit-default` options. It
could be superseded with the `MODULES_IMPLICIT_DEFAULT` environment
variable, that could be set with `config` module sub-command throught the
`implicit_default` option.

When implicit_default is disabled:
- module implicit default versions are excluded from 'module avail -d'
  results.
- false is returned when argument passed to the 'is-avail' sub-command
  is a directory without an explicit default set in it.
- and auto_handling is enabled, the requirement load mechanism will fail
  on requirement relying on module implicit default version

Fixes #179.