Add report_issue arg to getPathToModule proc
Add the report_issue argument to the getPathToModule procedure to
inhibit the report of issue raised during the search and resolution of a
passed module name.

Issue report is unwanted when performing a reverse lookup of the module
names resolving to a known modulefile (like done in procedures
getSimplifiedLoadedModuleList and getAllModuleResolvedName).

Issue report is also disabled in cmdModuleSearch, as found issues are
reported later on through the list returned by the getPathToModule call.