Force implicit_default enablement during 'restore'
When performing a 'module restore', forcibly enable the implicit_default
option to correctly match current state against the collection to
restore, even if this collection was saved with implicit_default enabled
and current setup at restore time has this option disabled.

implicit_default is then restored to its defined value once the match
is over. So the load of a module to restore from the collection may fail
when implicit_default is disabled if this module is an implicit default
and collection was saved with implicit_default enabled.