Rework no-indepth mod match in getModules
Match search query at corresponding search depth level when indepth mode
is disabled without implying that search query matches beginning of
module names. This is a preparatory work to a search match option where
query could be contained in module name and not only be the start of
module name.

As some results from findModules are removed through the process of
getModules, update element list of directories to fit results.