Add ability to control message verbosity
Add a ``verbosity`` configuration option to define the default message
verbosity level (between a ``normal`` and a ``debug`` level). This
option could be set at ``./configure`` time with ``--with-verbosity``
option. It could be superseded with the ``MODULES_VERBOSITY``
environment variable, that could be set with ``config`` module
sub-command through the ``verbosity`` option.

Defining a ``debug`` verbosity level at the command-line level is still
achieved by using the ``-D`` or ``--debug`` command-line switches.