Workaround for Tcsh history issue with wa_277 conf opt
Introduce the wa_277 configuration option to workaround an issue with Tcsh
history described on ticket #277.

When option is enabled (which sets the MODULES_WA_277 environment variable
to 1), an alternative module alias is defined which fixes the history
mechanism issue. However the alternative definition of the module alias
weakens shell evaluation of the code produced by modulefiles. Characters
with special meaning for Tcsh shell (like { and }) may not be used
anymore in shell alias definition elsewhere the evaluation of the code
produced by modulefiles will return a syntax error.

Fixes #277.