Rework ternary operator expr if res may equal 'nan'
Rework all the ternary operator expressions (eg. [expr {cond ? true :
false}] if their true or false result may be equal to the nan string
(whatever the case used to write it). When a 'nan' value is returned
from such an expression an error is raised, which breaks Modules as soon
as a modulefile, an alias or a symbolic version is named 'nan'.

These expressions are rewritten as bare if/then/else blocks.

Fixes #296.