Translate vers spec directly in modMatch
Pass the module version specification toward modMatch procedure to
translate it there for module matching. modMatch procedure is adapted to
apply the translation only if a 'trspec' argument is set to 1.

A pattern suffix 'psuf' argument is also added to give the ability to
append a suffix string to translated module name version.