Introduce range of version in mod version spec
A range of version may be specified after '@' with the help of ':'
character to express *greater or equal* (ex: `1.2:`), *lower or equal*
(ex: `:1.2`) or *between or equal* (ex: `1.1:1.2`) conditions.

Adapt procedures that parse module version specifier or compare module
version. Error is raised if version specified for range are not versions
or higher version is lower than lower version.

Non 'xdigit' character in higher part of version number is considered as
a non version number. It avoids symbols like `default`, `new` and so on
to get in the version range way.