Adapt parseModuleVersionSpecifier proc mlspec
Add a 'mlspec' argument to the parseModuleVersionSpecifier procedure to
indicate module specification comes or not from the 'ml' command. A
slightly different processing is applied in this case as module name may
be prefixed by a '-' character when specified over the 'ml' command to
indicate this module should be unloaded.

When mlspec is enabled two lists are returned by
parseModuleVersionSpecifier: one for the list of the module to load and
another for the modules to unload.