Fix -L/-d queries when upper dir not matching query
In case module avail query does not match a directory but only its
contained elements (for instance `module av mod/7` matches mod/7.1 and
mod/7.2 but not mod/), fix query processing to correctly return latest
or default element in case --latest/-L or --default/-d flags are set.

Add an ext_filter argument to getModules procedure to indicate that
directories of elements matching query should be part of the result list
even if these directories does not match query. This ext_filter argument
is enabled by listModules procedure when latest or default filters are
enabled. These directories are used when filtering is enabled to
determine what element should be displayed.