Introduce modEqStatic procedure
Introduce the modEqStatic procedure, a more static version of the modEq
procedure to use in repetitive comparison contexts like done in
getModules and modMatchAtDepth. In modEqStatic, search query is hard
wired which removes the need to call getModuleVersSpec each time the
procedure is called. Instead getModuleVersSpec is called when the
modEqStatic procedure is defined to get the hard wire elements. Doing so
saves a lot of time in the getModules routine.