Fix buffer overflow when working with large MODULEPATH
When MODULEPATH is larger than 1023 characters and holds environment
variable reference, a buffer overflow is triggered and leads to a
segmentation fault.

Correctly copying strings up to their length when resolving these
environment variable references fixes the issue.

Thanks to Dominique Martinet for all the help provided to this fix.

Fixes #311