Enclose names incl. space even if adv_vers_spec is disabled
Enclose module name and versions if it contains a space character even
if the advanced versio specifier mechanism is disabled.

Rework the advanced version specifier parsing done in
parseModuleVersionSpecifier procedure to also check argument passed if
advanced_version_spec mechanism is disabled and enclose them in curly
braces if a space character is found in them.