Fix unset-alias logic
The decision if "set-alias" should generate an actual alias or rather
a shell-function in output_set_alias worked differently then the one
for "unset-alias" in output_unset_alias.
This commit changes the latter to match the former, and also tries to
unset aliases AND functions if it's not known which of the two is
necessary (but only if this can be done without generating error
messages).