ts: compare sorted json output
As json output is not sorted by modulecmd.tcl it may happens that test
output found on some configurations or platforms is not the one
expected although it is correct.

To avoid these errors a new test procedure is added to compare stderr
output in a sorted manner, so error will be obtained if something is
missing in the result but not if the result is outputed in a different
order.