Ensure dir struct definition when only contains not-searched default
In case a directory has set for default element an hidden entry not
searched in this getModules processing, still record this directory in
dir_list with the default target referred in it (even if not part of
found_list result.

Doing so ensure that starting phase 3 of getModules processing, all
directories with a default defined will have an entry in found_list.