Record auto symbol in loaded environment
Once a module, which benefits from automatically defined symbolic
versions, is loaded, the associated auto symbols are recorded in loaded
environment among all the alternative names applying to the loaded
module (in `MODULES_LMALTNAME` environment variable).

Auto symbols are distinguished from regular symbols by the `as|` prefix
which is applied to them.