Refactor 'reportfd' state initialization
Transform the startPager procedure into initStateReportfd to have it
automatically called first time value of 'reportfd' state is queried (so
first time a message has to be outputed).

initStateReportfd procedure embeds the output of document start in case
of structured output mode. This output is done there to avoid an
infinite loop of reportfd state initialization.

Initialization of the 'reportfd' state is triggered in initErrorReport
if debug mode is enabled to have its reportDebug calls processed in
buffer mode to get processed in correct order (otherwise reportfd
initialization will appear first, before debug messages stored in
buffer).

'already_report' state is removed as it is now useless.