Format string and list for output
Add the strTo and listTo procedures to format a string or a list
of string making it compliant with the destination language. Add support
for Tcl and shell output, enclosing strings if they are empty or if they
contain a space character respectively with curly braces or single
quotes. And escaping those character if used in strings.

Use the new procedures in sh-to-mod to enclose/escape script name and
its argument to source in shell command. Also use listTo in reportCmd,
which was previously doing the same enclosing work.

Make use of reportCmd to output result of sh-to-mod sub-command, in
order to get the same formatted output than when doing a module display
(with tabs and colors if enabled).

Add a -nativeargrep argument to reportCmd procedure to use (when called
for sh-to-mod result) Tcl native string representation of list, which
automatically enclose string if needed and escape special characters if
present. Do not make this behavior a default as output made for module
display should not automatically enclose strings using variable value
reference (eg. $VAR).