Limit search to specified query in getPathToModule
Call getModules procedure with query specified to getPathToModule rather
on module root name. As a consequence getPathToModule recalls itself more
often when resolution hit a module alias or a symbolic version, as those
elements are less likely to be part of initial getModules result.

Delete the isSameModuleRoot procedure which is not used anymore.