No specific res for sym ver target in cmdModuleSearch
Modulefile targeted by symbolic version should be included in getModules
result along with symbolic version entry. So there is no need to
specifically resolve by default symbolic version target in
cmdModuleSearch procedure.

If symbol's target is found in error, record it if search query matches
symbol. And if symbol's target has not been found in directory where
symbol is defined, add this symbol target to the list of element to
resolve in other modulepath directories later on cmdModuleSearch
procedure.