Exclude hidden loaded mods from auto handling reports
Do not report hidden loaded modules in the informational messages of the
auto handling mechanisms unless these hidden modules have been
specifically loaded or unloaded by user, verbosity level is set to
debug or an higher level or any messages are recorded for the hidden
module evaluation.

The g_moduleHiddenEval array is introduced to record whether module
evaluation is hidden (hidden once loaded + automated evaluation of
module) or not.