Unset forbidden or set them as error elts in getModules res
Adapt `isModuleHidden` procedure and phase #2 of `getModules` procedure
to transform module entries set as forbidden in error entries
shipped with a *Access denied* message if these entries are specified in
search query or unset the forbidden entries not specified in search
query (like for hidden entries).

When search query matches forbidden modules, the error entry is
processed by procedure like `getPathToModule` to report the error
message to the user.