Add --before/--after options to module-{hide,forbid} cmds
Introduce the `--before` and `--after` options to `module-hide` and
`module-forbid` to make these commands apply only before a specific date
and/or after a specific date.

Procedure parseDateTimeArg is added to evaluate date time value
specified on the 2 new options and translate it into an epoch time for
comparison against current epoch time, saved in `clock_seconds` state.

Accepted datetime format are `YYYY-MM-DD` or `YYYY-MM-DDTHH:MM`.

Both options are ignored on Tcl<8.5 as -format option of clock scan
command is not available, which does not permit to work with selected
date time format.