Uncouple module hiding from module-forbid command
Do not imply hiding specified module on `module-forbid` command. Just
record module specification is said *forbidden* which will be checked to
disallow module load.

Doing so, hiding module is now only done through `module-hide` command
and the previous behavior of `module-forbid` command can now be achieved
with `module-hide --hard` + `module-forbid`.

This change also enables to forbid use of modules that are visible.

Hard-hiding level is lowered to *2*, as this spot was left empty through
this change.