unset env var when setting it to empty val on Win
When an empty string is set to an environment variable on Windows
platform, this variable is unset yet still seen defined by 'info exists'
command on Tcl, which leads to errors when trying to access variable
content. So when trying to set an empty string to an environment
variable on Windows platform, call for variable unset instead to reflect
underlying OS behavior.