Optimize hidden property search
Optimize internal recording of hidden module specification when parsing
modulerc files in order to reduce the time taken to test if a given
module is hidden or not.

Record hidden module specification in an array where key is module root
name. So it can be easy to check if there is some hidden specification
to test for a given module (based on root name). Within each array entry
hidden module specifications are sorted by hidden level to be able to
search these specifications from strongest to weakest hidding level.