Apply 'hidden' tag to module defined hidden
Set a `hidden` tag for each module targetted by a hidden definition
coming whether from a module-hide statement or by module name or version
starting with a dot character.

The `hidden` tag is not exported in MODULES_LMTAG for loading module as
the hidden state concerns the `avail` report only.