Introduce 'module-tag' modulefile command
Introduce the `module-tag` modulefile command to enable users to
associate their own tag to modulefiles. These tags are reported on
`avail` and `list` sub-command outputs.

This new command is intended to be used in modulerc evaluation context.
It is also recorded for modulefile evaluation context as it corresponds
to the evaluation context of the global and user RCs.