Update isModuleTagged to fetch info from collected tags
Add the collected argument to the isModuleTagged procedure to get the
ability to query tag information for designed modulefile from collected
tags data for module rather raw tag definition applying to module
specification.

Enable this query mode to determine if module is nearly-forbidden in
execute-modulefile procedure (as tags for evaluated modules are fetched
right before this test).

Disable 'nearly-forbidden' tag check when performing an unload
evaluation.