Apply tag sgr keys to element
When tags apply to a module to display, apply to this module name the
sgr keys that are defined to represent the tags. Each module tag
matching a sgr code is then graphically represented over module name and
not reported by its tag name or abbreviated name.

If an abbreviation exists for tag name, sgr key is matched against this
abbreviated string. If no abbreviation exists for a tag, sgr key is
matched against tag full name.

Currently only graphically render single tag.