Color tag by their own name
For tags set in tag_color_name configuration, apply graphical rendering
over their name rather than over the name of the module they are
attached to.

If an abbreviation is defined for tag, this abbreviation name should be
placed in tag_color_name configuration value to make it stand by itself
and not disappear over module name rendition.