Allow unload of sticky module if reloading
When reloading a module, during a `reload` or `restore` sub-command or
the processing of the *Dependency Reload* mechanism, allow the unload of
modules that are tagged sticky, as we know they will be reloaded.

Add `reloading_mod` state to record the name of modules that are
unloaded but that will be reloaded right after. With this new state,
unload of sticky modules can be allowed.