Swap sticky module if stickyness definition is preserved
Allow a sticky or super-sticky module to be swapped by another
modulefile version as soon as stickyness definition is still satisfied.
For instance, if a sticky tag is applied over *foo* module, *foo/1.0*
can be swapped by *foo/2.0* as *foo* stickyness is preserved. If the
sticky tag applies to *foo/1.0*, this module cannot be replaced by
another version of *foo* module.

Such swap may occur when processing a `restore` or a `switch`
sub-command.