Add term_width configuration
By default, term_width is equal to 0, but can be overridden
by the environment variable, which can be also overriden
by the command-line options -w or --width=.
When term_width is equal to 0, the output width is the full terminal
width, automatically detected by the module command.
If term_width is different from 0, this sets a specific number of
columns for the display of the output.

Closes #359