lib: add must_have_cookie arg to readFile proc
Update C-version of the readFile procedure to add it the
`must_have_cookie` argument that tells if the Modules magic cookie
(#%Module) has to be checked when reading file to abort reading as soon
as possible if read file must be a modulefile.

Bump version of Tcl Envmodules library to 1.3.0.

Fixes #375.