Build g_abbrevTag array in initConfTagAbbrev
In addition to build the g_tagAbbrev array in initConfTagAbbrev
procedure that associate each tag with their corresponding abbreviation,
the g_abbrevTag is now also built to associate each abbreviation with
their corresponding tag.

Some tags are known to share the same abbreviation, like hidden-loaded
and hidden tags, but these tags are used in different contexts. So when
building g_abbrevTag array skip tags not relevant to the current command
context