Add -o/--output command-line switches
Add the -o/--output command-line switches to supersedes the default
output configuration for avail and list sub-commands.

-o requires a value set as next argument whereas --output requires its
value to be joined to it after '=' character.

Adapt parseModuleCommandArgs, to be able to tell it to ignore errors on
first call pass performed on 'ml' command (as at the start module
sub-command is not known). Then a second call is made to raise the
errors for options not appropriate to the selected sub-command.

When the -o/--output options are used on an unsupported module
sub-command, or if these options are wrongly specified an error is
raised.