Validate {avail,list}_*output config values
Validate the value list set for the output content configuration based
on selected module sub-command. Error is raised as soon as an element in
value list is not defined in the allowed value list.

Add isDiffBetweenList procedure to check if an element in a list is not
part of a second list passed as argument.

Update getConf, cmdModuleConfig and parseModuleCommandArgs procedures to
correctly check the value passed to the output content configuration
options.