Update report_format state with terse/long mode
Change the report_format state to also hold the terse and long value in
case the --terse or --long mode is enabled. Change default value of
report_format state to regular. This way the state now indicates the
output mode which is expected with long and terse modes standing for a
specific plain output format.