Record variant specification in saved collection
When modules have been loaded with variants specified, record the
variants specification along module name and version when saving the
loaded environment as a collection.

In case the collection_pin_version configuration option is disabled,
only record in collection the variants whose value is not the default
one. This way the same behavior is obtained not to pin version or
variant if their default value is the one saved.