Record/use module designation including variant spec
Add the registerModuleDesignation and getModuleDesignation procedures to
report specified variants in addition to module name and version when
reporting load, unload or switch of modules.

Module name or variant specifications are enclosed between single quotes
if they contain a space character.

Module name and variant specifications are highlighted if configured
when registered to produce a graphically enhanced report message.

Variant specification is enclosed between {} to distinguish it from
module name and version.