Return updated module specification on getPathToModule
When getPathToModule is called to find a module specification from the
loaded module list, return an updated version of this module
specification fully matching the designation of found loaded module
(name, version and variants).

When unloading module, set this updated designation as the module
specified to unload. Which means querying the [module-info specified]
during an unload evaluation will return the loaded module specification
not the module specification set by user.