Set an optimized version of variant cmd for whatis eval
Add an optimized version of the variant modulefile command for the
whatis evaluation mode, not to specifically raise specification error
for this mode and barely set the variant to an empty string value in the
modulefile context. This way, operations on variant are lighter on this
global evaluation mode.

Variant definition errors still produce error which stops modulefile
evaluation but they are not reported unless if verbosity is set to debug
mode like any other errors during a whatis global evaluation.