Fix sh-to-mod to support fish 3.2
Update the sh-to-mod mechanism to support version 3.2 of the fish shell.
This new version introduces the `.` builtin command that should be
regexp-escaped when determining the shell functions or aliases defined
by the script analyzed by sh-to-mod.

An extra space character is also added by fish 3.2 at the end of shell
alias or function definition. Update parsing of these elements to ignore
such extra space.