Treat alt names of load{ed,ing} mods on modEq
Add the `ismodlo` argument to the modEq procedure to determine is passed
`mod` is a non-loaded module (0), a loaded module (1) or a loading
module (2). If `mod` is loaded or loading comparison against `pattern`
is also performed over the alternative names of mod.

This with one call to modEq, all module name and its alternatives are
tested against pattern: pattern properties are fetched and comparison
is prepared only once instead of calling modEq procedure over each
alternative name of module.