Add variant comparison ability to modEq
Adapt the modEq procedures to compare the variant specified for pattern
to the variant set for designated loaded or loading module. Only perform
such comparison if module name and version comparison is successful.

Variant comparison relies on the modVariantCmp procedure. Every variant
set for pattern should be found defined and with an identical value on
loaded/loading module. Otherwise there is no match.