Add modnamevr arg to execute-modulefile procedure
Add the modnamevr argument to the execute-modulefile procedure and save
this information into the modulenamevr state to make it available during
evaluation context.

Use that new modulenamevr state to pass variant list to the variant
procedure when module is unloading. This way the specifiedname on unload
mode is the specification set by user on command line or in modulefile
code, like on any other evaluation modes.