Record msgrecid in mod evaluation struct
Use the message recording id (msgrecid) to register the module
evaluation instead of the module name. Call for the module designation
based on this recording id when reporting evaluated modules. Update the
registerModuleEval and registerModuleEvalHidden procedures to use
msgrecid rather module name. In case of evaluation failure, pass the
module name with its specified variant to record for the
isModuleEvalFailed query.

When evaluated modules are reported, they use the full denomination
syntax with specified variants and space enclosed in single quotes.