Pass modnamevr to getUnmetDependentLoadedModuleList
Pass the module name and variants set as argument to the
getUnmetDependentLoadedModuleList procedure to make it use the variants
specified to determine the dependent modules. This information has to be
passed to the modEq procedure as no modulefile evaluation occurred yet
when getUnmetDependentLoadedModuleList is called, so variants have to be
transmitted as argument to modEq.