Add the 'editor' configuration option
Add the editor module configuration option to define the editor command
to use to interactively edit modulefiles. When specifically set, the
MODULES_EDITOR environment variable is defined. If not set, environment
variable VISUAL or EDITOR can be used to define the value of this
configuration (in this order).

Follow the same configuration scheme than Git with GIT_EDITOR, VISUAL
and EDITOR environment variables.