If evalid not found use mod arg as spec in getModuleDesignation
If the specified evaluation id passed to getModuleDesignation procedure
is not found, use the mod argument as a module specification to compute
the module designation, instead of using it as the resulting
designation.

Designation of switched-on module that fails to load is this way
formatted with the regular module designation syntax.