Compute list of alt and simplified names for loaded mods
Add the getLoadedAltAndSimplifiedName procedure that compute for
designated loaded module passed as argument the list of its alternative
names and simplified names (in case this module name version is the
implicit default).

This information is only fetched from the alternative names in case the
automated symbols are generated: simplified names are part of the
alternative names in this case. Otherwise the parent module name is
resolve against currently set modulepath to determine if parent module
resolves to currently loaded module version.

Add the ability on modEq procedure to compare a module pattern against
these alternative and simplified names of a loaded module: ismodlo
argument should be set to '3'.