Make restore process rely on regular modEq cmp
Adapt getMovementBetweenList procedure to compare element in provided
lists with the modEq procedure when lists are composed of module
specifications. Call modEq with a comparison of alternative and
simplified names of loaded modules.

The module restore process in cmdModuleRestore procedure only need one
call to getMovementBetweenList to get the list of module to unload, then
another call to get the list of module to load, in order to restore the
collection. Whether collection is version-pinned or not the modEq call
made during collection and loaded module comparison will also check
collection module name against loaded module simplified names.